Move tools to own package
