--- packages/app/src/tools/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "target": "es2017",
-    "lib": ["esnext"],
-    "allowJs": false,
-    "skipLibCheck": false,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noEmit": true,
-    "esModuleInterop": true,
-    "module": "CommonJS",
-    "moduleResolution": "node",
-    "typeRoots": ["../../node_modules/@types"],
-    "resolveJsonModule": true
-  },
-  "exclude": ["node_modules"],
-  "include": ["**/*.ts"]
-}
