+++ packages/tools/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@corona-dashboard/tools",
+  "private": true,
+  "version": "1.0.0",
+  "dependencies": {
+    "meow": "^8.0.0",
+    "ts-node": "^9.1.1"
+  },
+  "scripts": {
+    "build": "tsc",
+    "test": "echo 'No tests available'",
+    "test:ci": "echo 'No tests available'",
+    "format": "prettier --write **/*.{js,ts,tsx} *.{js,json,md,yml}",
+    "lint": "eslint .",
+    "type-check": "yarn compile",
+    "compile": "tsc --pretty --noEmit",
+    "validate-json": "ts-node src/validator/validate-all.ts",
+    "validate-single": "ts-node src/validator/validate-single.ts",
+    "generate-typescript": "ts-node src/generate-typescript.ts"
+  },
+  "devDependencies": {
+    "@types/meow": "^5.0.0"
+  }
+}
