+++ packages/tools/src/generate-typescript.ts
@@ -1,22 +1,30 @@
 /* eslint no-console: 0 */
 
-import fs from 'fs';
-import path from 'path';
-import { compile, JSONSchema } from 'json-schema-to-typescript';
-import { createValidateFunction } from '../validator/create-validate-function';
-import { getSchemaNames, schemaDirectory } from '../validator/get-schema-names';
+import fs from "fs";
+import path from "path";
+import { compile, JSONSchema } from "json-schema-to-typescript";
+import { createValidateFunction } from "./validator/create-validate-function";
+import { getSchemaNames, schemaDirectory } from "./validator/get-schema-names";
 
 // The directory where the resulting data.d.ts file will be saved
-const outputPath = path.join(__dirname, '..', '..', 'types');
+const outputPath = path.join(
+  __dirname,
+  "..", // tools
+  "..", // packages
+  "app",
+  "src",
+  "types"
+);
+
 const bannerComment =
   "/* tslint:disable */\n/**\n* This file was automatically generated by json-schema-to-typescript.\n* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,\n* and run 'yarn generate-typescript' to regenerate this file.\n*/\n\n";
 
-const schemaNames = getSchemaNames().filter((name) => name !== 'locale');
+const schemaNames = getSchemaNames().filter((name) => name !== "locale");
 
 const promises = schemaNames.map(generateTypeScriptFromSchema);
 
 Promise.all(promises).then((result) => {
-  saveDefinitionsFile(result.join('\n'));
+  saveDefinitionsFile(result.join("\n"));
 });
 
 /**
@@ -31,7 +39,7 @@ function generateTypeScriptFromSchema(schemaName: string) {
   const generateOptions = {
     cwd: path.join(schemaDirectory, schemaName),
     ignoreMinAndMaxItems: true,
-    bannerComment: '',
+    bannerComment: "",
   };
 
   return createValidateFunction(
@@ -51,9 +59,9 @@ function generateTypeScriptFromSchema(schemaName: string) {
 }
 
 function saveDefinitionsFile(typeDefinitions: string) {
-  const outputFile = path.join(outputPath, 'data.d.ts');
+  const outputFile = path.join(outputPath, "data.d.ts");
   fs.writeFileSync(outputFile, `${bannerComment}${typeDefinitions}`, {
-    encoding: 'utf8',
+    encoding: "utf8",
   });
   console.info(`Written typescript definitions output to file '${outputFile}'`);
 }
