+++ packages/tools/src/validator/ajv-keywords/equalsRootProperty.ts
