+++ packages/tools/src/validator/base-paths.ts
