+++ packages/tools/src/validator/custom-validations/validate-placeholders.ts
