+++ packages/tools/src/validator/get-schema-names.ts
