+++ packages/tools/src/validator/schema-information.ts
