+++ packages/tools/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "lib": ["esnext"],
+    "target": "ES2018",
+    "allowJs": false
+  },
+  "exclude": ["node_modules"],
+  "include": ["**/*.ts"]
+}
